修复单节点group by报错

+++ samples/spring-sharding-demo/src/main/java/com/easy/query/springshardingdemo/dto/OrderGroupWithSumOrderNoVO.java
@@ -12,6 +12,6 @@ import lombok.Data;
 @Data
 @EntityProxy
 public class OrderGroupWithSumOrderNoVO {
-    private Integer orderNoSum;
+    private Long orderNoSum;
     private String userId;
 }
